Allow users to set a trackpad/mouse preference to control wheel behaviour (#6755)
This change introduces a new preference of either `Trackpad` or `Mouse`
or no preference (in which case the `wheelBehavior` field of
`TLCameraOptions` is used) to control wheel behaviour.

For example, if selecting `Mouse` and using a mouse peripheral,
scrolling with the wheel will zoom in towards the cursor and if holding
command/control it pans the screen.

### New
<img width="434" height="663" alt="image"
src="https://github.com/user-attachments/assets/1ba3387e-3ebf-48f0-8bb8-b2a577baa2c4"
/>

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Test plan

- [x] Unit tests
- [ ] End to end tests

### Release notes

- Added preferences to set wheel behaviour for pointer peripherals

### API changes

- Added a new translation key
- Added a new preference to allow users to control wheel behaviour
`getInputMode`