add independent x and y self snapping to shapes (#6883)
This change introduces independent x and y line snapping for handles to
points on the same shape. Snapping already exists for individual points
but not for x and y lines.

The motivation for this change came from work on a cubic bezier shape,
where snapping control handles to start and end points on the x and y
axes aids in creating smooth curve shapes, as seen below.

![2025-10-06 at 13 59 53 - Violet
Ptarmigan](https://github.com/user-attachments/assets/68d60cd0-20af-4ac0-9f30-48a83aec4532)

Since users may not always want handles to snap to x and y lines, I've
introduced a `snapType` enum parameter to the `TLHandle` class that
allows to specify snap types `point` current behaviour and `align` new
behaviour. If it is not provided then we do not snap.

### Change type

- [ ] `bugfix`
- [x] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### API changes
Added a `snapType` property to the `TLHandle` class.

### Release notes

- Added a `snapType` which allows handles to snap to other snap points
on the x and y axis.