Add input element for local file picker to the DOM (#7037)
Apparently, for iOS, an
[undocumented](https://stackoverflow.com/questions/47664777/javascript-file-input-onchange-not-working-ios-safari-only)
requirement is for `input` elements to be added to the DOM for
`onchange` events to fire reliably.

We don't do this for uploading files locally, and therefore image
uploads fail silently and flakily for iOS users using Safari. Lets add
the `input` to the DOM. I also made sure that we resolve the promise if
we find no files, to avoid hanging promises.

### Change type

- [x] `bugfix`
- [ ] `improvement`
- [ ] `feature`
- [ ] `api`
- [ ] `other`

### Release notes

- Fixed a bug with iOS image uploads on Safari