Fix deleting (#7080)
I removed awaits since the editor recommended it, but it was fake news 😂

Looks like the editor incorrectly sees the methods as sync (`'await' has
no effect on the type of this expression.`, which is why I [removed them
in this PR](https://github.com/tldraw/tldraw/pull/7069). But it looks
like we still need them, or we don't get the most up to date state which
causes a [few issues when deleting
files](https://www.notion.so/tldraw/Deleting-files-is-glitchy-2a13e4c324c080d480b1c8f590b8133d?source=copy_link).

@ds300 maybe we should somehow fix these types?

### Change type

- [x] `bugfix`

<!-- CURSOR_SUMMARY -->
---

> [!NOTE]
> <sup>[Cursor Bugbot](https://cursor.com/dashboard?tab=bugbot) is
generating a summary for commit
6d2ee0f52be7e72a525698d039982a995c38f04e. Configure
[here](https://cursor.com/dashboard?tab=bugbot).</sup>
<!-- /CURSOR_SUMMARY -->